+++ src/main/consent-service/ConsentService.spec.ts
@@ -41,8 +41,9 @@ test('acceptConsent', () => {
     revoked: false
   };
   const serviceURL = `${serviceBaseUrl}/usage/accept`;
-  ConsentService.acceptConsent(option);
-  expect(httpSpy).toHaveBeenCalledWith(option, serviceURL, false, 'POST');
+  const headers = {requestId: 'requestId', lat: 'lat value', lon: 'lon value'}
+  ConsentService.acceptConsent(option, headers);
+  expect(httpSpy).toHaveBeenCalledWith(option, serviceURL, false, 'POST', undefined, headers);
 });
 
 test('getConsentVersionDetails', () => {
@@ -51,8 +52,9 @@ test('getConsentVersionDetails', () => {
     locale: 'locale'
   };
   const serviceURL = `${serviceBaseUrl}/versions/details/${option.consentid}?locale=${option.locale}`;
-  ConsentService.getConsentVersionDetails(option);
-  expect(httpSpy).toHaveBeenCalledWith(undefined, serviceURL, false, 'GET', option.access_token);
+  const headers = {requestId: 'requestId', lat: 'lat value', lon: 'lon value'}
+  ConsentService.getConsentVersionDetails(option, headers);
+  expect(httpSpy).toHaveBeenCalledWith(undefined, serviceURL, false, 'GET', option.access_token, headers);
 });
 
 test('acceptScopeConsent', () => {
@@ -62,8 +64,9 @@ test('acceptScopeConsent', () => {
     scopes: ['scopes']
   };
   const serviceURL = `${serviceBaseUrlV1}/scope/accept`;
-  ConsentService.acceptScopeConsent(option);
-  expect(httpSpy).toHaveBeenCalledWith(option, serviceURL, false, 'POST');
+  const headers = {requestId: 'requestId', lat: 'lat value', lon: 'lon value'}
+  ConsentService.acceptScopeConsent(option, headers);
+  expect(httpSpy).toHaveBeenCalledWith(option, serviceURL, false, 'POST', undefined, headers);
 });
 
 test('acceptClaimConsent', () => {
@@ -73,8 +76,9 @@ test('acceptClaimConsent', () => {
     accepted_claims: ['accepted_claims']
   };
   const serviceURL = `${serviceBaseUrlV1}/claim/accept`;
-  ConsentService.acceptClaimConsent(option);
-  expect(httpSpy).toHaveBeenCalledWith(option, serviceURL, false, 'POST');
+  const headers = {requestId: 'requestId', lat: 'lat value', lon: 'lon value'}
+  ConsentService.acceptClaimConsent(option, headers);
+  expect(httpSpy).toHaveBeenCalledWith(option, serviceURL, false, 'POST', undefined, headers);
 });
 
 test('revokeClaimConsent', () => {