+++ src/main/consent-service/ConsentService.ts
@@ -1,3 +1,4 @@
+import { HTTPRequestHeader } from "../common/Common.model";
 import { Helper } from "../common/Helper";
 import { AcceptClaimConsentRequest, AcceptConsentRequest, AcceptScopeConsentRequest, GetConsentDetailsRequest, GetConsentVersionDetailsRequest, RevokeClaimConsentRequest } from "./ConsentService.model";
 
@@ -33,6 +34,10 @@ export function getConsentDetails(options: GetConsentDetailsRequest) {
  *   consent_id: 'consent id',
  *   consent_version: 'consent version id',
  *   sub: 'masked sub'
+ * }, 
+ * {
+ *  lat: 'location latitude', 
+ *  lon: 'location longitude'
  * }).then((response) => {
  *   // the response will give you details of accepted consent.
  * }).catch((err) => {
@@ -40,9 +45,9 @@ export function getConsentDetails(options: GetConsentDetailsRequest) {
  * });
  * ```
  */
-export function acceptConsent(options: AcceptConsentRequest) {
+export function acceptConsent(options: AcceptConsentRequest, headers?: HTTPRequestHeader) {
   const _serviceURL = window.webAuthSettings.authority + "/consent-management-srv/v2/consent/usage/accept";
-  return Helper.createHttpPromise(options, _serviceURL, false, "POST");
+  return Helper.createHttpPromise(options, _serviceURL, false, "POST", undefined, headers);
 }
 
 /**
@@ -54,6 +59,10 @@ export function acceptConsent(options: AcceptConsentRequest) {
  *   consentid: 'your consent id',
  *   locale: 'browser accept language or custom language',
  *   access_token: 'your access token',
+ * }, 
+ * {
+ *  lat: 'location latitude', 
+ *  lon: 'location longitude'
  * }).then((response) => {
  *   // type your code here
  * }).catch((err) => {
@@ -61,13 +70,13 @@ export function acceptConsent(options: AcceptConsentRequest) {
  * });
  * ```
  */
-export function getConsentVersionDetails(options: GetConsentVersionDetailsRequest) {
+export function getConsentVersionDetails(options: GetConsentVersionDetailsRequest, headers?: HTTPRequestHeader) {
   let _serviceURL = window.webAuthSettings.authority + "/consent-management-srv/v2/consent/versions/details/" + options.consentid;
   if (options.locale) {
     _serviceURL += "?locale=" + options.locale;
   }
   // BREAKING TODO: remove access token as it is not needed to get consent version details
-  return Helper.createHttpPromise(undefined, _serviceURL, false, "GET", options.access_token);
+  return Helper.createHttpPromise(undefined, _serviceURL, false, "GET", options.access_token, headers);
 }
 
 /**
@@ -78,12 +87,16 @@ export function getConsentVersionDetails(options: GetConsentVersionDetailsReques
  *   client_id: 'your client id',
  *   sub: 'masked sub',
  *   scopes: [your scope consents]
+ *  }, 
+ *  {
+ *   lat: 'location latitude', 
+ *   lon: 'location longitude'
  *  });
  * ```
  */
-export function acceptScopeConsent(options: AcceptScopeConsentRequest) {
+export function acceptScopeConsent(options: AcceptScopeConsentRequest, headers?: HTTPRequestHeader) {
   const _serviceURL = window.webAuthSettings.authority + "/consent-management-srv/consent/scope/accept";
-  return Helper.createHttpPromise(options, _serviceURL, false, "POST");
+  return Helper.createHttpPromise(options, _serviceURL, false, "POST", undefined, headers);
 }
 
 /**
@@ -94,12 +107,16 @@ export function acceptScopeConsent(options: AcceptScopeConsentRequest) {
  *   client_id: 'your client id',
  *   sub: 'masked sub',
  *   accepted_claims: [your claim consents]
+ *  }, 
+ *  {
+ *   lat: 'location latitude', 
+ *   lon: 'location longitude'
  *  });
  * ```
  */
-export function acceptClaimConsent(options: AcceptClaimConsentRequest) {
+export function acceptClaimConsent(options: AcceptClaimConsentRequest, headers?: HTTPRequestHeader) {
   const _serviceURL = window.webAuthSettings.authority + "/consent-management-srv/consent/claim/accept";
-  return Helper.createHttpPromise(options, _serviceURL, false, "POST");
+  return Helper.createHttpPromise(options, _serviceURL, false, "POST", undefined, headers);
 }
 
 /**